Add Excptions nas classes services

+++ src/main/java/br/com/SmartMed/consultas/service/ConsultaService.java
@@ -1,6 +1,6 @@
 package br.com.SmartMed.consultas.service;
 
-import br.com.SmartMed.consultas.exception.ObjectNotFoundException;
+import br.com.SmartMed.consultas.exception.*;
 import br.com.SmartMed.consultas.model.ConsultaModel;
 import br.com.SmartMed.consultas.repository.ConsultaRepository;
 import br.com.SmartMed.consultas.rest.dto.ConsultaDTO;
@@ -18,30 +18,88 @@ public class ConsultaService {
     private ConsultaRepository consultaRepository;
 
     @Transactional(readOnly = true)
-    public ConsultaDTO buscaPorID(int id){
-        ConsultaModel consulta = consultaRepository.findById(id).orElseThrow(()-> new ObjectNotFoundException("Cpnsulta com o "+id+" Não encpntrado"));
+    public ConsultaDTO buscaPorID(int id) {
+        ConsultaModel consulta = consultaRepository.findById(id).orElseThrow(() -> new ObjectNotFoundException("Consulta com o " + id + " Não encpntrado"));
         return consulta.toDTO();
     }
 
     @Transactional(readOnly = true)
-    public List<ConsultaDTO> buscarTodos(){
+    public List<ConsultaDTO> buscarTodos() {
         List<ConsultaModel> consultas = consultaRepository.findAll();
         return consultas.stream().map(consulta -> consulta.toDTO()).collect(Collectors.toList());
     }
 
     @Transactional
-    public ConsultaDTO atualizar(ConsultaModel consulta){
-        return consultaRepository.save(consulta).toDTO();
+    public ConsultaDTO atualizar(ConsultaModel consulta) {
+        try {
+
+            if (!consultaRepository.existsById(consulta.getId())) {
+                throw new ObjectNotFoundException("Consulta  Não encpntrado");
+            }
+            return consultaRepository.save(consulta).toDTO();
+
+        } catch (DataIntegrityException e) {
+            throw new DataIntegrityException("Erro! Não foi possível atualizar a consulta " + consulta.getId() + " !");
+        } catch (ConstraintException e) {
+            if (e.getMessage() == null || e.getMessage().isBlank()) {
+                throw new ConstraintException("Erro ao atualizar a consulta " + consulta.getId() + ": Restrição de integridade de dados.");
+            }
+            throw e;
+        } catch (BusinessRuleException e) {
+            throw new BusinessRuleException("Erro ao atualizar a consulta " + consulta.getId() + "violação da regra de negocio");
+        } catch (SQLException e) {
+            throw new SQLException("Erro ao atualizar a consulta " + consulta.getId() + "Falha na conexão com o  banco de dados");
+        } catch (ObjectNotFoundException e) {
+            throw new ObjectNotFoundException("Erro ao atualizar a consulta " + consulta.getId() + "Não encontrado no bando de dados");
+        }
     }
 
     @Transactional
-    public ConsultaDTO salvar(ConsultaModel consulta){
-        return consultaRepository.save(consulta).toDTO();
+    public ConsultaDTO salvar(ConsultaModel consulta) {
+        try {
+            if (consultaRepository.existsById(consulta.getId())) {
+                throw new ObjectNotFoundException("Consulta já salvar");
+            }
+            return consultaRepository.save(consulta).toDTO();
+
+        } catch (DataIntegrityException e) {
+            throw new DataIntegrityException("Erro! Não foi possível salvar a consulta " + consulta.getId() + " !");
+        } catch (ConstraintException e) {
+            if (e.getMessage() == null || e.getMessage().isBlank()) {
+                throw new ConstraintException("Erro ao salvar a consulta " + consulta.getId() + ": Restrição de integridade de dados.");
+            }
+            throw e;
+        } catch (BusinessRuleException e) {
+            throw new BusinessRuleException("Erro ao salvar a consulta " + consulta.getId() + "violação da regra de negocio");
+        } catch (SQLException e) {
+            throw new SQLException("Erro ao salvar a consulta " + consulta.getId() + "Falha na conexão com o  banco de dados");
+        } catch (ObjectNotFoundException e) {
+            throw new ObjectNotFoundException("Erro ao salvar a consulta " + consulta.getId() + "Não encontrado no bando de dados");
+        }
     }
 
     @Transactional
-    public void deletar(ConsultaModel consulta){
-        consultaRepository.delete(consulta);
+    public void deletar(ConsultaModel consulta) {
+        try {
+            if (!consultaRepository.existsById(consulta.getId())) {
+                throw new ConstraintException("Consulta não existente");
+            }
+            consultaRepository.delete(consulta);
+
+        } catch (DataIntegrityException e) {
+            throw new DataIntegrityException("Erro! Não foi possível deletar a consulta " + consulta.getId() + " !");
+        } catch (ConstraintException e) {
+            if (e.getMessage() == null || e.getMessage().isBlank()) {
+                throw new ConstraintException("Erro ao deletar a consulta " + consulta.getId() + ": Restrição de integridade de dados.");
+            }
+            throw e;
+        } catch (BusinessRuleException e) {
+            throw new BusinessRuleException("Erro ao deletar a consulta " + consulta.getId() + "violação da regra de negocio");
+        } catch (SQLException e) {
+            throw new SQLException("Erro ao deletar a consulta " + consulta.getId() + "Falha na conexão com o  banco de dados");
+        } catch (ObjectNotFoundException e) {
+            throw new ObjectNotFoundException("Erro ao deletar a consulta " + consulta.getId() + "Não encontrado no bando de dados");
+        }
     }
 
 }
